feat: split up validate-messages per entrypoint for smarter bundling
